datathon-globo-recommendation/pyproject.toml
[tool.poetry]
name = "api-globo-news-recommendation"
version = "0.1.0"
description = ""
authors = ["gabriel-dantas98 <gabriel.dantasg98@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
scikit-learn = "^1.5.2"
mlflow = "^2.18.0"
torch = "^2.5.1"
yfinance = "^0.2.50"
numpy = "^2.1.3"
matplotlib = "^3.9.3"
joblib = "^1.4.2"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
opentelemetry-instrumentation-fastapi = "^0.49b2"
opentelemetry-instrumentation-logging = "^0.49b2"
opentelemetry-instrumentation-requests = "^0.49b2"
opentelemetry-exporter-otlp-proto-http = "^1.28.2"
prometheus-client = "^0.21.1"
pgvector = "^0.3.6"
transformers = "^4.49.0"
structlog = "^25.1.0"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
